requirements.txt
fastapi
uvicorn[standard]
pymupdf
pytesseract
Pillow
